fix(backend): parse Strava suffer_score as float, round to int
Strava's API returns suffer_score as a JSON float (e.g. 118.0) for some
activity types, but the raw struct declared it as int, breaking sync
with a JSON unmarshal error.

+++ backend/pkg/strava/client.go
@@ -151,7 +151,7 @@ type Activity struct {
 	AverageHeartrate   float64 `json:"average_heartrate"`
 	MaxHeartrate       float64 `json:"max_heartrate"`
 	AverageCadence     float64 `json:"average_cadence"`
-	SufferScore        int     `json:"suffer_score"`
+	SufferScore        float64 `json:"suffer_score"`
 }
 
 // GetActivities fetches recent activities for the authenticated athlete
